refactoring stats and add geolocalization
