+++ stats/IP2LOCATION-LITE-DB1/README_LITE.TXT
@@ -1,0 +1,20 @@
+------
+README
+------
+
+IP2Location LITE - http://lite.ip2location.com
+
+Copyright (c) 2001-2018 Hexasoft Development Sdn. Bhd.
+All Rights Reserved.
+
+1. IP2Location LITE Edition is free package with accuracy up to Class C (192.168.1.X) only.
+
+2. IP2Location provides commercial packages with accuracy up to all IP addresses in Class C (192.168.1.1). You can get the commercial editions from http://www.ip2location.com.
+
+3. Licensees may copy, distribute, display and perform the work and make derivative works based on IP2Location LITE only if they give the IP2LOCATION.COM credits in the manner specified below.
+
+All sites, advertising materials and documentation mentioning features or use of this database must display the following acknowledgment:
+
+"This site or product includes IP2Location LITE data available from http://www.ip2location.com."
+
+IP2Location is a registered trademark of Hexasoft Development Sdn Bhd. All other trademarks are the property of their respective owners.
