+++ stats/caddy-logs/readme.txt
@@ -1,0 +1,1 @@
+This folder contains Caddy's logs.
