+++ stats/parsed-logs/readme.txt
@@ -1,0 +1,1 @@
+This folder contains logs parsed by EnerMaps stats module.
