+++ stats/pg-logs/readme.txt
@@ -1,0 +1,1 @@
+This folder contains PostgresSQL's logs.
