+++ stats/requirements.txt
@@ -1,1 +1,4 @@
 pandas==1.3.0
+sqlalchemy==1.3.20
+psycopg2-binary==2.8.6
+python-dateutil==2.8.2
